Use json binder when parse form in new project

+++ server/entities/repository/repository.go
@@ -40,10 +40,10 @@ func FindByGithubRepoID(id int64) (*Repository, error) {
 }
 
 // CreateRepository create repository record based on github repository
-func CreateRepository(id int, oauthToken string) (*Repository, error) {
+func CreateRepository(id int64, oauthToken string) (*Repository, error) {
 	// confirm github
 	h := hub.New(oauthToken)
-	githubRepo, err := h.GetRepository(id)
+	githubRepo, err := h.GetRepository(int(id))
 	if err != nil {
 		return nil, err
 	}
